Clipboard Test
Sadly, it's not possible to put something in the clipboard with Chrome when the user issue an "copy" event, which is kinda anticlimatic.

In other browsers, copy only fires when there is something selected, our "virtualization" of an input is not considered "something".

We have no choice but to give up on Copy+Paste for inputs. Such is life.