Gui Input Main Class This huge class proved to be a real pain to create, lots of small details an input has to process.
Four hundred lines for an Input UI simulation in canvas. This obviously counts as "reinventing the wheel", which is looked down upon, but it's *absolutely* perfect for numeric inputs, e-mails and whatnot, the only exception is that it is missing copy, paste and cut, that just couldn't be implemented. Also, the graphical part cannot be interpreted differently between browsers.

Anyways, this is an experiment, I can afford to put my time in learning how silly things like inputs works, so this TextInput in Canvas taught me a lot of cool things about how things work.

Some small tricks here and there, need to put a bit of work on undo/redo, but overall I'm very happy with this class!

Worst con of using this virtualization of Text Input: *Absolutely no support for mobile devices*. Tough luck, not my applications main focus thought, so I'll set that aside.