Redo / Undo / Paste to Gui-Input
Created a way to redo and undo on the text input, capable of memorizing 10 states.

Used the paste event to put foreing text into the inputs, while it's impossible to do the contrary, this will help. I still need to check if the copied text is a valid number or not.